Move remaining actions from actions.ts to actions/general.ts
